+++ showhive-domain/domain/src/main/java/com/showhive/performance/domain/Performance.java
@@ -1,7 +1,6 @@
 package com.showhive.performance.domain;
 
 import com.showhive.BaseEntity;
-import com.showhive.ShowHiveException;
 import com.showhive.performance.exception.PerformanceErrorCode;
 import com.showhive.performance.exception.PerformanceException;
 import jakarta.persistence.Column;
@@ -9,8 +8,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 import lombok.AccessLevel;
